VLBIImagePriors

Documentation for VLBIImagePriors.

VLBIImagePriors.AngleTransform

VLBIImagePriors.CenteredImage

VLBIImagePriors.DiagonalVonMises

VLBIImagePriors.GMRFCache

VLBIImagePriors.GMRFCache

VLBIImagePriors.GaussMarkovRF

VLBIImagePriors.GaussMarkovRF

VLBIImagePriors.GaussMarkovRF

VLBIImagePriors.ImageDirichlet

VLBIImagePriors.ImageSimplex

VLBIImagePriors.ImageSphericalUniform

VLBIImagePriors.ImageUniform

VLBIImagePriors.SphericalUnitVector

VLBIImagePriors.WrappedUniform

VLBIImagePriors.alr

VLBIImagePriors.alr!

VLBIImagePriors.alrinv

VLBIImagePriors.AngleTransform — Type

AngleTransform

A transformation that moves two vector x and y to an angle θ. Note that is x and y are normally distributed then the resulting distribution in θ is uniform on the circle.

VLBIImagePriors.CenteredImage — Type

CenteredImage(x, y, σ, p)

Regularizes a general image prior p such that the center of light is close the the origin of the imag. After regularization the log density of the prior is modified to

\[ \log p(I) \to \log p(I) - \frac{(x_C^2 + y_C^2)^2}{2\sigma\^2} N_x N_y\]

where N_x and N_y are the number of pixels in the x and y direction of the image, and $x_C, y_C$ are the center of light of the image I.

VLBIImagePriors.DiagonalVonMises — Type

DiagonalVonMises(μ::Real, κ::Real)
DiagonalVonMises(μ::AbstractVector{<:Real}, κ::AbstractVector{<:Real})

Constructs a Von Mises distribution, with mean μ and concentraion parameter κ. If μ and κ are vectors then this constructs a independent multivariate Von Mises distribution.

Notes

This is a custom implementation since the version in Distributions.jl has certain properties that do not play well (having an support only between [-π+μ, π+μ]) with usual VLBI problems. Additionally this distribution has a special overloaded product_distribution method so that concatenating multiple DiagonalVonMises together preserves the type. This is helpful for Zygote autodiff.

VLBIImagePriors.GMRFCache — Type

struct GMRFCache{A, TD, M}

A cache for a Gaussian Markov random field. This cache is useful when defining a hierarchical prior for the GRMHD.

Fields

Λ

Intrinsic Gaussian Random Field pseudo-precison matrix. This is similar to the TSV regularizer

D

Gaussian Random Field diagonal precision matrix. This is similar to the L2-regularizer

λQ

The eigenvalues of the Λ matrix which is needed to compute the log-normalization constant of the GMRF.

Examples

julia> mean = zeros(2,2) # define a zero mean
julia> cache = GRMFCache(mean)
julia> prior_map(x) = GaussMarkovRF(mean, x[1], x[2], cache)
julia> d = HierarchicalPrior(prior_map, product_distribution([Uniform(-5.0,5.0), Uniform(0.0, 10.0)]))
julia> x0 = rand(d)

VLBIImagePriors.GMRFCache — Method

GMRFCache(mean::AbstractMatrix)

Contructs the GMRFCache from the mean image mean. This is useful for hierarchical priors where you change the hyperparameters of the GaussMarkovRF, λ and Σ.

VLBIImagePriors.GaussMarkovRF — Type

struct GaussMarkovRF{T, M<:AbstractArray{T, 2}, P, C, TDi} <: Distributions.Distribution{Distributions.Matrixvariate, Distributions.Continuous}

A image prior based off of the first-order Gaussian Markov random field. This is similar to the combination of L₂ and TSV regularization and is equal to

λ TSV(I-M) + Σ⁻¹L₂(I-M) + lognorm(λ, Σ)

where λ and Σ are given below and M is the mean image and lognorm(λ,Σ) is the log-normalization of the random field and is needed to jointly infer I and the hyperparameters λ, Σ.

Fields

m

The mean image of the Gaussian Markov random field

λ

The pixel correlation function of the random field. This is the TSV hyperparameter and controls the pixel-to-pixel correlation. As λ→0 this become a white-noise process.

Σ

The variance of each pixel of the random field. As Σ→∞ we revert to a pure TSV prior, which is an improper or intrinsic Gaussian Markov random field.

cache

The internal cache of the process. See GMRFCache for more information

dims

The dimensions of the image.

Examples

julia> mimg = zeros(6, 6) # The mean image
julia> λ, Σ = 2.0, 1.0
julia> d = GaussMarkovRF(mimg, λ, Σ)
julia> cache = GMRFCache(mimg) # now instead construct the cache
julia> d2 = GaussMarkovRF(mimg, λ, Σ, cache)
julia> invcov(d) ≈ invcov(d2)
true

VLBIImagePriors.GaussMarkovRF — Method

GaussMarkovRF(mean::AbstractMatrix, λ, Σ, cache::GMRFCache)

Constructs a first order Gaussian Markov random field with mean image mean and correlation λ and diagonal covariance Σ and the precomputed GMRFCache cache.

VLBIImagePriors.GaussMarkovRF — Method

GaussMarkovRF(mean::AbstractMatrix, λ, Σ)

Constructs a first order Gaussian Markov random field with mean image mean and correlation λ and diagonal covariance Σ.

VLBIImagePriors.ImageDirichlet — Type

ImageDirichlet(α::AbstractMatrix)
ImageDirichlet(α::Real, ny, nx)

A Dirichlet distribution defined on a matrix. Samples from this produce matrices whose elements sum to unity. This is a useful image prior when you want to separately constrain the flux. The α parameter defines the usual Dirichlet concentration amount.

Notes

Much of this code was taken from Distributions.jl and it's Dirichlet distribution. However, some changes were made to make it faster. Additionally, we use define a custom rrule to speed up derivatives.

VLBIImagePriors.ImageSimplex — Type

ImageSimplex(ny,nx)

This defines a transformation from ℝⁿ⁻¹ to the n probability simplex defined on an matrix with dimension ny×nx. This is a more natural transformation for rasterized images, which are most naturally represented as a matrix.

Notes

Much of this code was inspired by TransformVariables. However, we have specified custom rrules using Enzyme as a backend. This allowed the simplex transform to be used with Zygote and we achieved an order of magnitude speedup when computing the pullback of the simplex transform.

VLBIImagePriors.ImageSphericalUniform — Type

ImageSphericalUniform(nx, ny)

Construct a distribution where each image pixel is a 3-sphere uniform variable. This is useful for polarization where the stokes parameters are parameterized on the 3-sphere.

Currently we use a struct of vectors memory layout. That is the image is described by three matrices (X,Y,Z) grouped together as a tuple, where each matrix is one direction on the sphere, and we require norm((X,Y,Z)) == 1.

VLBIImagePriors.ImageUniform — Type

ImageUniform(a::Real, b::Real, nx, ny)

A uniform distribution in image pixels where a/b are the lower/upper bound for the interval. This then concatenates ny×nx uniform distributions together.

VLBIImagePriors.SphericalUnitVector — Type

SphericalUnitVector{N}()

A transformation from a set of N+1 vectors to the N sphere. The set of N+1 vectors are inherently assumed to be N+1 a distributed according to a unit multivariate Normal distribution.

Notes

For more information about this transformation see the Stan manual. In the future this may be depricated when is merged.

VLBIImagePriors.WrappedUniform — Type

WrappedUniform(period)

Constructs a potentially multivariate uniform distribution that is wrapped a given period. That is

d = WrappedUniform(period)
logpdf(d, x) ≈ logpdf(d, x+period)

for any x.

If period is a vector this creates a multivariate independent wrapped uniform distribution.

VLBIImagePriors.alr! — Method

alr!(x, y)

Computes the additive logit transform inplace. This is useful for moving from a normal distribution to the logistic normal, which lives on the simplex.

Notes

This function is mainly to transform the GaussMarkovRF to live on the simplex. In order to preserve the nice properties of the GRMF namely the log det we only use y[begin:end-1] elements and the last one is not included in the transform. This shouldn't be a problem since the additional parameter is just a dummy in that case and the Gaussian prior should ensure it is easy to sample from.

VLBIImagePriors.alr — Method

alr(y)

Computes the additive logit transform of the parameter y. This is useful for moving from a normal distribution to the logistic normal, which lives on the simplex.

```julia-repl julia> y = randn(10) julia> x = alr(y) julia> sum(x) ≈ 1 true

Notes

This function is mainly to transform the GaussMarkovRF to live on the simplex. In order to preserve the nice properties of the GRMF namely the log det we only use y[begin:end-1] elements and the last one is not included in the transform. This shouldn't be a problem since the additional parameter is just a dummy in that case and the Gaussian prior should ensure it is easy to sample from.

VLBIImagePriors.alrinv — Method

alrinv(x)

Compute the inverse alr transform. Note that we fill the last element with zero.